Announce message schemas to the ground; wkt event msgs take natural ids
NamedMsg now requires postcard_schema::Schema, PortSchema::Postcard carries
the payload's OwnedNamedType, and the downlink announces one SetMsgMetadata
per distinct tapped packet id on every connect (the db already persists
them), so ground tools can decode any message channel generically. The ten
wkt event messages (LogEvent, alarm set, sequence set, ReloadSequences,
WiringManifest) drop their hand-held [224,x] ids for the blanket impl's
schema-name-hash ids — historical records under the old ids stay cleanly
unmatched, and the pin tests now guard against type renames re-keying
recordings. AlarmTarget.element_index becomes u64 (postcard-schema has no
usize). ABI v9 -> v10; python dist pins follow. Adds State::msg_log_iter for
enumerating recorded channels.

+++ libs/metor-panel/src/alarms/mod.rs
@@ -183,7 +183,7 @@ impl AlarmState {
     fn targets(def: &AlarmDef, component_id: ComponentId, element: usize) -> bool {
         def.target.as_ref().is_some_and(|target| {
             target.component_id == component_id
-                && target.element_index.map(|i| i == element).unwrap_or(true)
+                && target.element_index.map(|i| i as usize == element).unwrap_or(true)
         })
     }
